Add Firebase configuration and dependencies
Integrate Firebase into the Android project by adding the necessary
configuration files and dependencies. This includes the addition of
`google-services.json`, updates to `build.gradle` and `settings.gradle`,
and the creation of `firebase_options.dart`. The `firebase_core`
dependency is also added to `pubspec.yaml` to enable Firebase
functionality in the app.